consolidate summarization examples
